+++ src/com/intellij/idea/plugin/hybris/codeInspection/rule/impex/ImpexNoUniqueValueInspection.kt
@@ -1,6 +1,6 @@
 /*
  * This file is part of "SAP Commerce Developers Toolset" plugin for IntelliJ IDEA.
- * Copyright (C) 2019-2024 EPAM Systems <hybrisideaplugin@epam.com> and contributors
+ * Copyright (C) 2019-2025 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -52,7 +52,7 @@ private class NoUniqueValueVisitor(private val problemsHolder: ProblemsHolder) :
                         .associate { (name, attrs) ->
                             name to attrs
                                 .flatMap { it.valueGroups }
-                                .mapNotNull { it.value }
+                                .map { it.value }
                         }
 
                     if (distinctCommonAttrsNames.isEmpty()) {